database id changes, case comment changes, profile changes

+++ Hippra/Models/DTO/AddHistoryLogDto.cs
@@ -5,7 +5,7 @@ namespace Hippra.Models.DTO
     public class AddHistoryLogDto
     {
         public HistoryLogType Type { get; set; }
-        public int PostID { get; set; }
+        public long PostID { get; set; }
         public long CommentId { get; set; }
 
         public string UserId { get; set; }
