+++ Hippra/Models/SQL/CaseLike.cs
@@ -3,8 +3,8 @@ namespace Hippra.Models.SQL
 {
     public class CaseLike
     {
-        public int ID { get; set; }
-        public int CaseId { get; set; }
+        public long ID { get; set; }
+        public long CaseId { get; set; }
         public string LikedByUserId { get; set; }
         public DateTime LikeDate { get; set; }
